refactor(table): move transaction.rs into lance-table
Building a manifest from a transaction reads and writes only table metadata.
The file has no dependency on Dataset, Session, DataFusion, or async I/O
beyond restore_old_manifest, which uses only ObjectStore and CommitHandler,
so it belongs at the table layer.

The file moves unchanged apart from import paths. lance::dataset::transaction
becomes a re-export, so no caller in lance, the Python bindings, or the Java
bindings changes.

build_manifest, restore_old_manifest, modifies_same_metadata and
upsert_key_conflict widen from pub(crate) to pub, since their callers in
io/commit.rs and io/commit/conflict_resolver.rs are now in another crate.
lance's own public API is unchanged.

The split into a module tree is deliberately left for a follow-up so this
step stays a single detected rename.
